Fixed cache response issues
Firstly, fixed response by converting the cache data (which is in a list) into a string, and encoding it. Secondly I had to make sure variable cacheData was set to the unencoded cacheData as a string instead of list (before being encoded) so it would print properly in the skeleton code without erroring.